+++ CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-## [0.1.3] - 2026-08-12
+## [0.1.3] - 2026-08-15
 
 ### Added